pull common code out
